infrastructure/requirements.txt
selenium>=4.0.0
webdriver-manager>=3.0.0
pyyaml>=6.0
google-generativeai>=0.3.0
beautifulsoup4>=4.11.1
requests>=2.28.1